% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_BHRejVecClass.R
\name{BHRejVec}
\alias{BHRejVec}
\title{Produce a BHRejVec object}
\usage{
BHRejVec(FDR = 0.1, testset, ifF = FALSE, eps = 0.5)
}
\arguments{
\item{FDR}{A vector of numeric values between 0 and 1.}

\item{testset}{A samVec object containing data and necessary information.}

\item{ifF}{Test type. When ifF == TRUE, standard F tests are used and; otherwise,
equivalence F tests.}

\item{eps}{A numeric value specifying the equivalence test tolerance. Only valid when
ifF == FALSE.}
}
\value{
A BHRejVec object is returned. A BHRejVec is a list contains
\describe{
 \item{BHProc}{A list of BHRej objects with different FDR values}
 \item{labels}{The labels indicating groups specified in the samVec object.}
 \item{dataType}{A string the same with the dataType in the samVec object}
 \item{testType}{A string of "DE" or "EE" based on ifF or not.}
 \item{FDR}{A numeric vector containing the FDR values used.}
}
}
\description{
\code{BHRejVec} defines BHRejVec class and produce a BHRejVec object. A BHRejVec object
is a combination of BHRej objects of different FDR values. Also, a samVec object
is used to construct a BHRejVec object.
}
\examples{
data(GCwPADataA)
t1 = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"),
 dataType = "Example")
DE = BHRejVec(FDR = c(0.1, 0.01), testset = t1)
EE = BHRejVec(FDR = c(0.1, 0.01), testset = t1, ifF = TRUE)
}

