% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_order}
\alias{Cpp_order}
\title{The order of a vector}
\usage{
Cpp_order(vec)
}
\arguments{
\item{vec}{A vector of numerics.}
}
\value{
A vector of integers containing the order the given vector is returned.
}
\description{
\code{Cpp_order} returns the order of a given vector. It has the same purpose as 
\code{\link[base]{order}}, but is implemented with Cpp and much faster.
}

