% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA_PriorFuns.R
\name{Cpp_p_function}
\alias{Cpp_p_function}
\title{Calculate the p-value}
\usage{
Cpp_p_function(epsilon = 0, ifF = FALSE, myData = list())
}
\arguments{
\item{epsilon}{A value used for equivalence tests. Only valid when ifF == FALSE.}

\item{ifF}{A bool value indicating test type. When ifF = TRUE, standard F test is used; otherwise, equivalence F test.}

\item{myData}{A list of vector of numerics.}
}
\value{
P value.
}
\description{
\code{p_function} calculate the p-value of data given the epsilon and the test type.
\code{Cpp_p_function}, an alternative for the same purpose, is much faster and recommended.
}
\examples{
a = rnorm(100)
b = rnorm(150)
d = rnorm(200)
e = runif(250)
p_function(0, FALSE, list(a, b, d, e))
Cpp_p_function(0, FALSE, list(a, b, d, e))
}

