% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_rank}
\alias{Cpp_rank}
\title{The rank of a vector}
\usage{
Cpp_rank(vec)
}
\arguments{
\item{vec}{A vector of numerics.}
}
\value{
A vector of integers containing the rank the given vector is returned.
}
\description{
\code{Cpp_rank} returns the order of a given vector. It has the same purpose as
\code{\link[base]{rank}}, but is implemented with Cpp and much faster.
}

