% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEE.R
\docType{package}
\name{DEEE}
\alias{DEEE}
\alias{DEEE-package}
\title{Provide tools for DE and EE analysis.}
\description{
The DEEE package provdies different graphical and analytical tools, such as BHRej, 
HeatMap, etc., to implement Bonjamini-Hochberg procedure, weave plots, heapmap, PCA
analysis, etc.
}

