% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM_HeatnAxisnLeg.R
\name{DEEEAxis}
\alias{DEEEAxis}
\title{Add axis}
\usage{
DEEEAxis(lm = 0, rm, tm, bm, sep = numeric(),
  labels = letters[1:(length(sep) + 1)], colSet = NULL, bdPar = list(ifbd
  = TRUE, col = "white"), txtPar = NULL)
}
\arguments{
\item{lm}{A numeric indicating left margin.}

\item{rm}{A numeric indicating right margin.}

\item{tm}{A numeric indicating top margin.}

\item{bm}{A numeric indicating bottom margin.}

\item{sep}{A vector of numeric used as delimiter of axis columns.}

\item{labels}{A vector of strings added to each separate part.}

\item{colSet}{A vector of strings specifying the colors added to each separate part.}

\item{bdPar}{A list of parameters specifying the borders of axis.}

\item{txtPar}{A list of parameters specifying the labels in the axis.}
}
\description{
\code{DEEEAxis} adds axis to existing heapmaps.
}
\examples{
testMatrix = matrix(sample(1:100), 10, 10)
colnames(testMatrix) = 1:10
rownames(testMatrix) = 1:10
DEEEHeat(testMatrix)
DEEEAxis(lm = 0, rm = 10, bm = 10, tm = 11, sep = c(1, 3, 5), colSet = 2:5)
}

