% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM_HeatnAxisnLeg.R
\name{DEEEHeat}
\alias{DEEEHeat}
\title{Core function for heatmap.}
\usage{
DEEEHeat(mydata, mycol = NULL, graPar = list(), textNum = NULL,
  textX = NULL, textY = NULL, Seg = TRUE)
}
\arguments{
\item{mydata}{A matrix containing data.}

\item{mycol}{A vector of strings indicating color range.}

\item{graPar}{A list of global graphics parameters.}

\item{textNum}{A list specifying texts plotted in each cell of the heatmap.
NA for no number.}

\item{textX}{A list containing parameters specifing X-axis. NA for no plotting.}

\item{textY}{A list containing parameters specifing Y-axis. NA for no plotting.}

\item{Seg}{A bool value indicating whether plot separating lines or not.}
}
\value{
A vector of numerics indicating the border of the plot, (xleft, xright, ybottom,
ytop) is returned.
}
\description{
\code{DEEEHeat} is the basic function to construct a heatmap. This function plot the
core heat map and
return the coordinates containing the border of the plot.
}
\examples{
testMatrix = matrix(sample(1:100), 10, 10)
colnames(testMatrix) = 1:10
rownames(testMatrix) = 1:10
DEEEHeat(testMatrix)
}

