% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM_HeatnAxisnLeg.R
\name{DEEELegend}
\alias{DEEELegend}
\title{Add legends}
\usage{
DEEELegend(pos, textPos = NULL, cex = 1, labels, colSet = NULL)
}
\arguments{
\item{pos}{A vector of length four specifying (xleft, ybottom, xright, ytop),
indicating the position of legend.}

\item{textPos}{A list of two vectors of numerics indicating the position of the label
positions.}

\item{cex}{A numeric indicating the cex of text in legend.}

\item{labels}{A vector of strings added to the legend.}

\item{colSet}{A vector of strings giving the color range of legend. Usually not need to
be specified.}
}
\description{
\code{DEEELegend} adds legends to existing heatmap.
}
\examples{
testMatrix = matrix(sample(1:100), 10, 10)
colnames(testMatrix) = 1:10
rownames(testMatrix) = 1:10
DEEEHeat(testMatrix)
DEEEAxis(lm = 0, rm = 10, bm = 10, tm = 11, sep = c(1, 3, 5), colSet = 2:5)
DEEELegend(pos = c(10.5, 5, 11, 10), labels = c(0, 50, 100), textPos = list(x = c(11.5,
 11.5, 11.5), y = c(5, 7.5, 10)))
}

