% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocData.R
\docType{data}
\name{GCwPADataA}
\alias{GCwPADataA}
\title{Preprocessed MAQC AFX data.}
\format{A matrix with 16797 rows and 30 columns.}
\usage{
GCwPADataA
}
\description{
A MAQC AFX dataset preprocessed with GCRMA and filtered with the P/M/A call in 
MAS5.0 packages. Only genes from pool A are picked.
}
\keyword{datasets}

