% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WV_weaveClass.R
\name{medWV}
\alias{medWV}
\title{Produce a weave object with restrictions.}
\usage{
medWV(SVobj, degree = 4)
}
\arguments{
\item{SVobj}{A samVec object containing data and other information.}

\item{degree}{A integer specifying the degree of the polynomail logistic regression that
applied to fit the data.}
}
\value{
A weave object is returned, which is a list containing
\describe{
 \item{fitY}{A list of lists of numerics containing the fitted response values.}
 \item{rankit}{A list of numerics containing the rankit of the means of the data matrix.}
 \item{order}{A list of integers containing the order the rankit.}
 \item{nWV}{An integers indicating the numbers of weave.}
 \item{interval}{c(0, 1).}
 \item{Function}{"mad".}
 \item{ifeq}{">".}
 \item{selCol}{A list of vector of integers containing absolute indexes of choosen
 columns.}
 \item{colInd}{A list of vector of integers containing relative indexes of columns.}
 \item{labels}{A vector of characters, integers or short string marking the groups.}
 \item{dataType}{A long string containing useful information of the data.}
}
}
\description{
\code{medWV} defines and produces a weave object with FUN = ``median'',
interval = c(0, 1) and ifeq = ``>''. See also \code{\link{weave}}, which produces a weave
object with no restriction but is much slower. See \code{\link{plot.weave}}
for plotting instructions.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"),
 dataType = "Example")
m1 = medWV(SVobj = testset, degree = 6)
plot(m1, legPar = list(cex = 1, ncol = 1), mainPar = list(main = "Example"))
}

