% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA_PriorFuns.R
\name{paraMerge}
\alias{paraMerge}
\title{Merge two lists}
\usage{
paraMerge(oldP, newP)
}
\arguments{
\item{oldP}{A list containing the old parameters.}

\item{newP}{A list containing the new parameters.}
}
\value{
A list of merged parameters is returned.
}
\description{
\code{paraMerge} merges two arguments list into a new one, in which the old parameters
are over-written by new parameters.
}
\examples{
paraMerge(list(t1 = 1, t2 = 2, t3 = 3, t4 = 4), list(t3 = 4, t4 = 5, t5 = 6, t6 = 7))
}

