% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_PCA.R
\name{PCA}
\alias{PCA}
\title{Produce a PCA object.}
\usage{
PCA(SVobj, scale = F)
}
\arguments{
\item{SVobj}{A samVec object containing data and other information.}

\item{scale}{A bool value indicating whether scaling is used or not?}
}
\value{
A PCA object is returned, which is list containing:
\describe{
 \item{pcx1}{A matrix containing the first principle conponent.}
 \item{pcx2}{A matrix containing the second principle conponent.}
 \item{colInd}{A list of vector of integers containing relative indexes of columns.}
 \item{labels}{A vector of characters, integers or short string marking the groups.}
 \item{dataType}{A long string containing useful information of the data.}
 \item{nGroup}{An interger indicating the number of groups specified by selCol.}
 \item{nCol}{An integer indicating the number of columns of data.}
 \item{nRow}{A integers indicating the number of rows of data.}
}
}
\description{
\code{PCA} defines and produces a PCA object. See \code{\link{plot.PCA}}
for plotting instructions.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"))
pca1 = PCA(testset)
plot(pca1, mainPar = list(labels = "Example", y = 1.55), pcaPar = list(cex = 1, font = 2))
}

