% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_BHRejVecClass.R
\name{PIE.BHRejVec}
\alias{PIE.BHRejVec}
\title{PIE plot BHRejVec class.}
\usage{
\method{PIE}{BHRejVec}(obj, objEE, figcol = 1, graPar = list(),
  mainPar = list(), innerPar = list(), innerPIE = list(),
  legPar = list(), xshrink = 1, yshrink = 0.95, ...)
}
\arguments{
\item{obj, objEE}{Objects of BHRejVec class to be ploted.}

\item{figcol}{An integer indicating the number of columns of figures plotted.}

\item{graPar}{A list of parameters adjusting global graphics setting.}

\item{mainPar}{A list of parameter adjusting the main title.}

\item{innerPar}{A list of parameters adjusting the inner pie plots setting.}

\item{innerPIE}{A list of parameters adjusting the pie plots setting.}

\item{legPar}{A list of parameters adjusting legend.}

\item{xshrink, yshrink}{Numeric values between 0 and 1 adjusting the height of figures and
width.}

\item{...}{ignored}
}
\description{
\code{PIE.BHRejVec} produces pie plots given two BHRejVec objects. THey must be produced
with the same samVec objec and have the same FDR values but different test types.
}
\examples{
data(GCwPADataA)
t1 = samVec(GCwPADataA, selCol = list(1:5, 6:10, 11:15, 16:20, 21:25, 26:30), labels = LETTERS[1:6])
EE = BHRejVec(FDR = c(0.1, 0.05, 0.01, 0.005, 0.001), SVobj = t1)
DE = BHRejVec(FDR = c(0.1, 0.05, 0.01, 0.005, 0.001), SVobj = t1, ifF = TRUE)
pdf("test.pdf", width = 12, height = 8)
PIE(DE, EE, figcol = 3, mainPar = list(y = 1.5, main = "TITLE", cex = 2.5),
 innerPIE = list(radius = 1, col = c("blue", "red", "green")),
 innerPar = list(mar = c(1.1, 2.1, 6.1, 0)), legPar = list(x = "bottomright", cex = 1.5))
dev.off()
}

