% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WV_grpWVClass.R
\name{plot.grpWV}
\alias{plot.grpWV}
\title{Plot method for grpWV class}
\usage{
\method{plot}{grpWV}(x, indSet = NULL, legPar = list(cex = 0.6, ncol = 3),
  graPar = list(), mainPar = list(), colSet = NULL, ltySet = NULL, ...)
}
\arguments{
\item{x}{A grpWV object.}

\item{indSet}{A vector of integers indicating which weaves are plotted.}

\item{legPar}{A list of legend parameters.}

\item{graPar}{A list of global graphics parameters.}

\item{mainPar}{A list of background (including main title) parameters.}

\item{colSet, ltySet}{Vectors of numerics or characters specifying col types or line}

\item{...}{ignored}
}
\description{
\code{grpWV} is the plot method for grpWV class.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"))
madw1 = madWV(SVobj = testset, degree = 6)
plot(madw1, legPar = list(cex = 1, ncol = 1), mainPar = list(main = "Example"))
}

