% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_PCA.R
\name{plot.PCA}
\alias{plot.PCA}
\title{The plot method for a PCA object.}
\usage{
\method{plot}{PCA}(x, mainPar = list(), graPar = list(), pcaPar = list(),
  ...)
}
\arguments{
\item{x}{A PCA object.}

\item{mainPar}{A list of parameters specifying the main title.}

\item{graPar}{A list of global graphics parameters.}

\item{pcaPar}{A list of parameters adjusting the PCA points.}

\item{...}{ignored}
}
\description{
\code{plot.PCA} plots all the points given their first and second components. Different
points in different groups are given different colors and labels.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"))
pca1 = PCA(testset)
plot(pca1, mainPar = list(labels = "Example", y = 1.55), pcaPar = list(cex = 1, font = 2))
}

