% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_samVecClass.R
\name{plot.samVec}
\alias{plot.samVec}
\title{Plot methods of class samVec}
\usage{
\method{plot}{samVec}(obj, graPar = NULL, col = 1:obj$nGroup, lty = rep(1,
  obj$nGroup), main = "KDE", legPar = list())
}
\arguments{
\item{obj}{A samVec object.}

\item{graPar}{A list of parameters adjusting the global graphics.}

\item{col}{A vector of strings or integers specifying colors.}

\item{lty}{A vector of integers specifying line types.}

\item{main}{A string specifying main title.}

\item{legPar}{A list of parameters specifying the legend.}
}
\description{
\code{plot.samVec} divides the data by their groups and calcutes their densities
separately. Their densities are plotted together in the same figure.
}

