% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WV_weaveClass.R
\name{plot.weave}
\alias{plot.weave}
\title{Plot method for weave class}
\usage{
\method{plot}{weave}(x, indSet = NULL, legPar = list(cex = 0.6, ncol = 3),
  graPar = list(), mainPar = list(), ltySet = NULL, colSet = NULL, ...)
}
\arguments{
\item{x}{A weave object.}

\item{indSet}{A vector of integers indicating which weaves are plotted.}

\item{legPar}{A list of legend parameters.}

\item{graPar}{A list of global graphics parameters.}

\item{mainPar}{A list of background (including main title) parameters.}

\item{colSet, ltySet}{Vectors of numerics or characters specifying col types or line types.}

\item{...}{ignored}
}
\description{
\code{plot.weave} is the plot method for weave class.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"))
m1 = medWV(SVobj = testset, degree = 6)
plot(m1, legPar = list(cex = 1, ncol = 1), mainPar = list(main = "Example"))
}

