% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA_PriorFuns.R
\name{plotMat}
\alias{plotMat}
\title{Calcute the layout coordinates}
\usage{
plotMat(nfig, figcol, xshrink = 1, yshrink = 1)
}
\arguments{
\item{nfig}{An integer indicating the total number of figures.}

\item{figcol}{An integer indicating the columns of figures.}

\item{xshrink, yshrink}{The right or top shrink.}
}
\value{
A list of positions is returned with each element containing a numeric vector
of length 4, standing for (xleft, xright, ybottom, ytop)
}
\description{
\code{plotMat} calcutes the layout coordinates given numbers of figures and columns of figures.
The order should be from top to bottom and left to right.
}
\examples{
plotMat(10, 3)
}

