% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_samVecClass.R
\name{samVec}
\alias{samVec}
\title{Produce a samVec object}
\usage{
samVec(myDataset, selCol = NULL, labels = NULL, dataType = "MAQC")
}
\arguments{
\item{myDataset}{A matrix.}

\item{selCol}{A list of vector of integers containing absolute indexes of choosen columns.}

\item{labels}{A vector of characters or short strings marking the groups. The length of
labels should be the same with the length of selCol. If not specifying, a vector of
integer is given.}

\item{dataType}{A string contains data type or other useful imformation.}
}
\value{
\code{samVec} returns a samVec object. A samVec is a list contains:
\describe{
 \item{data}{A matrix containing the data imported by myDataset.}
 \item{nCol}{An integer indicating the number of columns of data.}
 \item{nGroup}{An interger indicating the number of groups specified by selCol.}
 \item{colInd}{A list of vector of integers containing relative indexes of columns.}
 \item{selCol}{A list of vector of integers containing absolute indexes of choosen columns.}
 \item{nRow}{A integers indicating the number of rows of data.}
 \item{labels}{A vector of characters, integers or short string marking the groups.}
 \item{dataType}{A long string containing useful information of the data.}
}
}
\description{
\code{samVec} defines the samVec object given necessary data and information.
The objects of samVec class are the base for further analysis.
}
\examples{
datamatrix = matrix(rnorm(100000), 1000, 100)
testset = samVec(datamatrix, selCol = list(1:5, 11:15, 21:25), dataType = "Example")
plot(testset, main = testset$dataType)
}

