% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WV_weaveClass.R
\name{weave}
\alias{weave}
\title{Produce a weave object.}
\usage{
weave(dataset, degree = 4, FUN = "median", interval = c(0, 1),
  ifeq = FALSE)
}
\arguments{
\item{dataset}{A samVec object containing data and other information.}

\item{degree}{A integer specifying the degree of the polynomail logistic regression that
applied to fit the data.}

\item{FUN}{A string specifying the function used in weave.}

\item{interval}{A value between 0 and 1 to restrict the rankit used.}

\item{ifeq}{An bool value indicating whether >= or > is used in comparison. This argument
is only valid when it comes tied values.}
}
\value{
A weave object is returned, which is a list containing
\describe{
 \item{fitY}{A list of lists of numerics containing the fitted response values.}
 \item{rankit}{A list of numerics containing the rankit of the means of the data matrix.}
 \item{order}{A list of integers containing the order the rankit.}
 \item{nWV}{An integers indicating the numbers of weave.}
 \item{interval}{An interval indicating the range of rankit that is used.}
 \item{Function}{A string indicating the function used in weave.}
 \item{ifeq}{An bool value indicating whether >= or > is used in comparison. This
  argument is only valid when it comes tied values.}
 \item{selCol}{A list of vector of integers containing absolute indexes of choosen
 columns.}
 \item{colInd}{A list of vector of integers containing relative indexes of columns.}
 \item{labels}{A vector of characters, integers or short string marking the groups.}
 \item{dataType}{A long string containing useful information of the data.}
}
}
\description{
\code{weave} defines and produces a weave object. The difference between weave and
\code{\link{grpWV}} is their way to apply FUN. In weave, FUN is applied to every
single column; while in grpWV, FUN takes the whole columns in a group as argument.
See also \code{\link{medWV}}, in which a weave object also is produced, but with
FUN = ``median'', interval = c(0, 1) and ifeq = ``>''. See \code{\link{plot.weave}}
for plotting instructions.
}
\examples{
data(GCwPADataA)
testset = samVec(GCwPADataA, selCol = list(1:5, 11:15, 21:25), labels = c("E", "R", "T"),
 dataType = "Example")
medw1 = weave(dataset = testset, degree = 6, FUN = "median", interval = c(0, 1))
plot(medw1, legPar = list(cex = 1, ncol = 1), plotPar = list(main = "Example"))
}

