% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_BHRejClass.R
\name{PIE.BHRej}
\alias{PIE.BHRej}
\title{PIE plot}
\usage{
\method{PIE}{BHRej}(DE, EE, graPar = NULL, main = NULL, piePar = list())
}
\arguments{
\item{DE, EE}{Objects of BHRej class.}

\item{graPar}{A list of parameters adjusting global graphic.}

\item{main}{A string specifying title.}

\item{piePar}{A list of parameters adjusting the pie plot.}
}
\description{
\code{PIE.BHRej} produces a pie plot for objects of BHRej class. The first argument
contains imformation of DE and the second EE. They should be produced with the same data
 set, but different test types.
}
\examples{
DE = BHRej(FDR = 0.1, runif(1000, 0, 0.1))
EE = BHRej(FDR = 0.1, runif(1000, 0, 0.1))
PIE(DE = DE, EE = EE, piePar = list(radius = 0.9, main = "What"))
}

