% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_BHRejClass.R
\name{PIE}
\alias{PIE}
\title{Generic function of PIE plot}
\usage{
PIE(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Others parameters.}
}
\description{
\code{PIE} produces a pie plot. More details could be found in \code{\link{PIE.BHRej}}
and \code{\link{PIE.BHRejVec}}.
}

