% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_samVecClass.R
\name{plot.samVec}
\alias{plot.samVec}
\title{Plot methods of class samVec}
\usage{
\method{plot}{samVec}(x, graPar = NULL, colSet = 1:x$nGroup,
  ltySet = rep(1, x$nGroup), main = "KDE", legPar = list(), ...)
}
\arguments{
\item{x}{A samVec object.}

\item{graPar}{A list of parameters adjusting the global graphics.}

\item{colSet}{A vector of strings or integers specifying colors.}

\item{ltySet}{A vector of integers specifying line types.}

\item{main}{A string specifying main title.}

\item{legPar}{A list of parameters specifying the legend.}

\item{...}{ignored}
}
\description{
\code{plot.samVec} divides the data by their groups and calcutes their densities
separately. Their densities are plotted together in the same figure.
}
\examples{
datamatrix = matrix(rnorm(100000), 1000, 100)
testset = samVec(datamatrix, selCol = list(1:5, 11:15, 21:25), dataType = "Example")
plot(testset, main = testset$dataType)
}

